Moved refreshParametersTable to be private function, try catch for postgresmodel.refresh
