+++ extensions/arc/src/ui/dashboards/postgres/postgresCoordinatorNodeParametersPage.ts
@@ -7,7 +7,7 @@ import * as azdata from 'azdata';
 import * as loc from '../../../localizedConstants';
 import { IconPathHelper } from '../../../constants';
 import { PostgresParametersPage } from './postgresParameters';
-import { PostgresModel } from '../../../models/postgresModel';
+import { PostgresModel, EngineSettingsModel } from '../../../models/postgresModel';
 
 export class PostgresCoordinatorNodeParametersPage extends PostgresParametersPage {
 
@@ -29,6 +29,10 @@ export class PostgresCoordinatorNodeParametersPage extends PostgresParametersPag
 
 	protected get description(): string {
 		return loc.coordinatorNodeParametersDescription;
+	}
+
+	protected get engineSettings(): EngineSettingsModel[] {
+		return this._postgresModel.coordinatorNodeEngineSettings;
 	}
 
 	protected async saveParameterEdits(): Promise<void> {
@@ -62,21 +66,5 @@ export class PostgresCoordinatorNodeParametersPage extends PostgresParametersPag
 				this._postgresModel.controllerModel.azdataAdditionalEnvVars,
 				session);
 		*/
-	}
-
-	protected refreshParametersTable(): void {
-		this._parameters = this._postgresModel.coordinatorNodeEngineSettings.map(engineSetting => this.createParameterComponents(engineSetting));
-
-		this._parametersTable.data = this._parameters.map(p => {
-			if (p.information) {
-				// Container to hold input component and information bubble
-				const valueContainer = this.modelView.modelBuilder.flexContainer().withLayout({ alignItems: 'center' }).component();
-				valueContainer.addItem(p.valueComponent, { CSSStyles: { 'margin-right': '0px' } });
-				valueContainer.addItem(p.information, { CSSStyles: { 'margin-left': '5px' } });
-				return [p.parameterName, valueContainer, p.description, p.resetButton];
-			} else {
-				return [p.parameterName, p.valueComponent, p.description, p.resetButton];
-			}
-		});
 	}
 }
