+++ extensions/arc/src/ui/dashboards/postgres/postgresWorkerNodeParametersPage.ts
@@ -8,7 +8,7 @@ import * as azdataExt from 'azdata-ext';
 import * as loc from '../../../localizedConstants';
 import { IconPathHelper } from '../../../constants';
 import { PostgresParametersPage } from './postgresParameters';
-import { PostgresModel } from '../../../models/postgresModel';
+import { PostgresModel, EngineSettingsModel } from '../../../models/postgresModel';
 
 export class PostgresWorkerNodeParametersPage extends PostgresParametersPage {
 
@@ -33,6 +33,10 @@ export class PostgresWorkerNodeParametersPage extends PostgresParametersPage {
 	protected get description(): string {
 		// TODO update to loc.workerNodesParametersDescription
 		return loc.nodeParametersDescription;
+	}
+
+	protected get engineSettings(): EngineSettingsModel[] {
+		return this._postgresModel.workerNodesEngineSettings;
 	}
 
 	protected async saveParameterEdits(engineSettings: string, session: azdataExt.AzdataSession): Promise<void> {
@@ -60,21 +64,5 @@ export class PostgresWorkerNodeParametersPage extends PostgresParametersPage {
 			this._postgresModel.engineVersion,
 			this._postgresModel.controllerModel.azdataAdditionalEnvVars,
 			session);
-	}
-
-	protected refreshParametersTable(): void {
-		this._parameters = this._postgresModel.workerNodesEngineSettings.map(engineSetting => this.createParameterComponents(engineSetting));
-
-		this._parametersTable.data = this._parameters.map(p => {
-			if (p.information) {
-				// Container to hold input component and information bubble
-				const valueContainer = this.modelView.modelBuilder.flexContainer().withLayout({ alignItems: 'center' }).component();
-				valueContainer.addItem(p.valueComponent, { CSSStyles: { 'margin-right': '0px' } });
-				valueContainer.addItem(p.information, { CSSStyles: { 'margin-left': '5px' } });
-				return [p.parameterName, valueContainer, p.description, p.resetButton];
-			} else {
-				return [p.parameterName, p.valueComponent, p.description, p.resetButton];
-			}
-		});
 	}
 }
